Update parse-command.ts
The issue seems to be related to how the bot is parsing the command from the message. The current implementation might be expecting the command to be at the start of the message.

i modified the parseCommand function in parse-command.ts to search the entire message for the command, not just at the start. i used a regular expression to find the command and its arguments within the message.